pyproject.toml
[project]
name = "warden"
version = "0.1.0"
description = "AI Agent for Continuous Codebase Vigilance"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0",
    "sqlalchemy>=2.0",
    "gitpython>=3.1",
    "pyyaml>=6.0",
]

[project.scripts]
warden = "warden.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
